clinker-core: move JoinSide and AggregateStrategy into the plan layer
These two compile-time selections are produced by the planner and merely
consumed by the lower runtime layers, but their definitions lived down in
`executor::combine` (`JoinSide`) and `aggregation::error`
(`AggregateStrategy`). That forced `plan/` to reach upward into the
executor and aggregation modules to name types it actually owns.

Introduce `plan::types` as the home for both enums and repoint every
caller — the planner, the combine and aggregate executor arms, the
aggregation engine, the lowering pass, and the join-resolution tests — at
the new location. The plan layer no longer imports either enum from a
higher module, removing a circular-looking dependency edge ahead of the
eventual crate split.

`JoinSide` keeps its derives and `AggregateStrategy` keeps its
snake_case serde representation; only their module path changed.
